ProgressBarをTopFragmentに移動、ボタン連打防止
